requirements.txt
langchain-community
langchain-google-genai
sentence-transformers
streamlit
chromadb
python-dotenv
google-generativeai
matplotlib
seaborn
langchain_mcp_adapters
langgraph
mcp
streamlit